Refactor puzzle solver tests
Rename TestCases to Data, as in AutoFixture.xUnit3 refactor.
Change data to 2 values instead of tuple, but still define as tuple.

Rename IPuzzleSolverTests to IPuzzleSolutionData, that's all it does.
No longer require its implementation used in PuzzleSolverTests to
inherit PuzzleSolverTests, implementation is enough.
Makes it much simpler to make tests for multiple solver implementations
for the same puzzle later on if needed.